Fix some Clippy warnings
* clippy::unnecessary_cast warnings
* clippy::needless_borrow
* clippy::suspicious_to_owned .  This was an actual performance bug.
  The to_owned() was a data copy that we didn't need.

+++ src/libxfuse/dir3.rs
@@ -218,7 +218,7 @@ impl Dir2Data {
         start_block: u64,
     ) -> Dir2Data {
         let offset = start_block * (superblock.sb_blocksize as u64);
-        buf_reader.seek(SeekFrom::Start(offset as u64)).unwrap();
+        buf_reader.seek(SeekFrom::Start(offset)).unwrap();
 
         let hdr = Dir3DataHdr::from(buf_reader.by_ref());
